[memprof] Rename RawMemProfReader.{cpp,h} to MemProfReader.{cpp,h} (NFC)
This patch renames RawMemProfReader.{cpp,h} to MemProfReader.{cpp,h},
respectively.  Also, it re-creates RawMemProfReader.h just to include
MemProfReader.h for compatibility with out-of-tree users.
